# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API for TBOX Backend.",
        "title": "TBOX Backend API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/login": {
            "post": {
                "description": "User can login in system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Login with phone number",
                "parameters": [
                    {
                        "type": "string",
                        "description": "0987654321",
                        "name": "phone_number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/common.DataFormat"
                        }
                    },
                    "400": {
                        "description": "something went wrong",
                        "schema": {
                            "$ref": "#/definitions/common.DataFormat"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.DataFormat"
                        }
                    }
                }
            }
        },
        "/api/v1/ping": {
            "get": {
                "description": "Ping health check server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Ping server",
                "responses": {
                    "200": {
                        "description": "pong",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/resend-otp": {
            "post": {
                "description": "User resend otp if user not receive otp before",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Resend OTP to phone number of user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "9c070de4-d6aa-4d53-b9f3-02eacb5a1d05",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/common.DataFormat"
                        }
                    },
                    "400": {
                        "description": "something went wrong",
                        "schema": {
                            "$ref": "#/definitions/common.DataFormat"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.DataFormat"
                        }
                    }
                }
            }
        },
        "/api/v1/verify": {
            "post": {
                "description": "User verify phone number before accept system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Verify phone number of user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "9c070de4-d6aa-4d53-b9f3-02eacb5a1d05",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "123456",
                        "name": "otp",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/common.DataFormat"
                        }
                    },
                    "400": {
                        "description": "something went wrong",
                        "schema": {
                            "$ref": "#/definitions/common.DataFormat"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.DataFormat"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.DataFormat": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}